refactor: polishing code
