--- REFACTORING_SUMMARY.md
@@ -1,153 +1,0 @@
-# MicroMouse Simulator - Refactoring Summary
-
-## Changes Made
-
-### 1. **Fixed Step Generation Issue** ✅
-   - **Problem**: The first grid state wasn't being displayed during step-by-step maze generation
-   - **Root Cause**: The grid was only updated after generation completed, not after initialization or each step
-   - **Solution**: 
-     - Grid is now updated immediately after `initialize()` is called
-     - Grid is updated after each `step()` call during animation
-     - This ensures the UI can display the complete generation process from the very first state
-
-### 2. **Separated GUI Code from Main** ✅
-   - **Problem**: Main.cpp contained all simulation logic and GUI code, violating separation of concerns
-   - **Solution**: Created a new `Simulator` class that encapsulates all simulation state and logic
-   
-   **New Files Created:**
-   - `include/Simulator.hpp` - Simulator class header
-   - `src/Simulator.cpp` - Simulator class implementation
-   
-   **Main.cpp Changes:**
-   - Reduced from 282 lines to ~65 lines
-   - Now only handles:
-     - Window creation
-     - Main event loop
-     - Delegating to Simulator
-   - All simulation logic moved to Simulator class
-
-### 3. **Code Organization Improvements**
-   - **Simulator Class** encapsulates:
-     - Maze generation state
-     - Solver state
-     - Animation state
-     - UI instance
-     - All simulation logic methods
-   
-   - **Main.cpp** now only handles:
-     - Application entry point
-     - Window management
-     - Main loop coordination
-
-### 4. **Clang-Tidy Compliance** ✅
-   All changes follow clang-tidy requirements:
-   - Proper naming conventions (lower_case for functions/variables)
-   - No non-const global variables (state encapsulated in Simulator class)
-   - Modern C++ practices (auto, range-based loops where appropriate)
-   - Proper const-correctness
-
-## Technical Details
-
-### Step Generation Fix
-
-**Before:**
-```cpp
-auto start_generation(bool step_by_step) -> void {
-    maze_gen.initialize(rows, cols);
-    is_generating = true;
-    // Grid not available until generation completes!
-}
-```
-
-**After:**
-```cpp
-auto Simulator::start_generation(bool step_by_step) -> void {
-    maze_gen_.initialize(maze_rows_, maze_cols_);
-    
-    // FIX: Update grid immediately so first state is visible
-    maze_grid_ = maze_gen_.get_grid();
-    
-    is_generating_ = true;
-    // ...
-}
-```
-
-**Animation Loop Fix:**
-```cpp
-if (is_generating_) {
-    const bool still_running = maze_gen_.step();
-    // Update grid after EACH step so UI can display changes
-    maze_grid_ = maze_gen_.get_grid();
-    
-    if (!still_running) {
-        // Convert to graph when done
-        maze_graph_ = MazeSolver::convert_to_graph(...);
-        is_generating_ = false;
-    }
-}
-```
-
-### Architecture Improvement
-
-**Old Architecture:**
-```
-Main.cpp (282 lines)
-├── Global SimulationState struct
-├── run_solver() function
-├── start_generation() function
-├── reset_simulator() function
-├── main() with all event handling and animation logic
-└── UI class (separate)
-```
-
-**New Architecture:**
-```
-Main.cpp (65 lines)
-└── main() - minimal, just window and loop
-
-Simulator class
-├── All simulation state (private members)
-├── initialize()
-├── handle_event()
-├── update()
-├── render()
-├── update_mouse_position()
-└── Private helper methods
-    ├── run_solver()
-    ├── start_generation()
-    ├── reset_simulator()
-    └── handle_instant_solve_click()
-```
-
-## Testing
-
-Added comprehensive tests in `tests/MazeGen-Test.cpp`:
-- `InitializationShowsFirstGrid` - Verifies grid is available after init
-- `StepByStepGeneration` - Verifies grid updates after each step
-
-All tests pass ✅
-
-## Build System
-
-Updated `CMakeLists.txt`:
-- Added `src/Simulator.cpp` to `CORE_SOURCES`
-- No other changes needed
-
-## Benefits
-
-1. **Better Separation of Concerns**: Main.cpp is now minimal and focused
-2. **Easier to Test**: Simulator class can be tested independently
-3. **Better Maintainability**: Related code is grouped together
-4. **Fixed Bug**: Step generation now shows all states correctly
-5. **Clang-Tidy Compliant**: Follows all coding standards
-
-## Files Modified
-
-- `src/Main.cpp` - Simplified to use Simulator class
-- `CMakeLists.txt` - Added Simulator.cpp to build
-- `tests/MazeGen-Test.cpp` - Added comprehensive tests
-
-## Files Created
-
-- `include/Simulator.hpp` - Simulator class header
-- `src/Simulator.cpp` - Simulator implementation
