--- include/solvers/BFSSolver.hpp
@@ -1,39 +1,0 @@
-#ifndef MICRO_MOUSE_BFS_SOLVER_HPP
-#define MICRO_MOUSE_BFS_SOLVER_HPP
-
-#include "SolverInterface.hpp"
-#include <queue>
-#include <vector>
-
-class BFSSolver : public SolverInterface {
-public:
-    void initialize(const std::vector<Cell>& grid, 
-                    const Graph& graph, 
-                    size_t start, 
-                    size_t end, 
-                    size_t rows, 
-                    size_t cols) override;
-
-    bool step() override;
-    std::vector<size_t> get_path() const override;
-    std::vector<size_t> get_visited_order() const override;
-    bool is_solved() const override;
-
-private:
-    const Graph* graph_ = nullptr;
-    size_t start_node_ = 0;
-    size_t end_node_ = 0;
-    size_t total_nodes_ = 0;
-
-    std::queue<size_t> work_queue_;
-    std::vector<bool> visited_;
-    std::vector<size_t> predecessors_;
-    
-    // For visualization
-    std::vector<size_t> visited_order_;
-    
-    bool solved_ = false;
-    bool finished_ = false; // True if queue empty or target found
-};
-
-#endif // MICRO_MOUSE_BFS_SOLVER_HPP
