+++ src/Main.cpp
@@ -11,17 +11,13 @@
 #include <iostream>
 
 // Application constants
-static constexpr int K_WINDOW_WIDTH  = 1280;
-static constexpr int K_WINDOW_HEIGHT = 720;
-static constexpr int K_MIN_WINDOW_WIDTH  = 800;
-static constexpr int K_MIN_WINDOW_HEIGHT = 600;
+static constexpr int WINDOW_WIDTH  = 1280;
+static constexpr int WINDOW_HEIGHT = 720;
+static constexpr int MIN_WINDOW_WIDTH  = 800;
+static constexpr int MIN_WINDOW_HEIGHT = 600;
 
-/**
- * @brief Application entry point.
- * @return 0 on successful exit, 1 on error.
- */
 auto main() -> int {
-    sf::RenderWindow window(sf::VideoMode(K_WINDOW_WIDTH, K_WINDOW_HEIGHT),
+    sf::RenderWindow window(sf::VideoMode(WINDOW_WIDTH, WINDOW_HEIGHT),
                             "MicroMouse Simulator Pro",
                             sf::Style::Default);
     window.setFramerateLimit(60);
@@ -53,28 +49,28 @@ auto main() -> int {
                     unsigned int width = event.size.width;
                     unsigned int height = event.size.height;
                     
-                    if (width < K_MIN_WINDOW_WIDTH || height < K_MIN_WINDOW_HEIGHT) {
-                        width = std::max(width, static_cast<unsigned int>(K_MIN_WINDOW_WIDTH));
-                        height = std::max(height, static_cast<unsigned int>(K_MIN_WINDOW_HEIGHT));
+                    if (width < MIN_WINDOW_WIDTH || height < MIN_WINDOW_HEIGHT) {
+                        width = std::max(width, static_cast<unsigned int>(MIN_WINDOW_WIDTH));
+                        height = std::max(height, static_cast<unsigned int>(MIN_WINDOW_HEIGHT));
                         window.setSize(sf::Vector2u(width, height));
                     }
                     
                     // Update the view to match the new window size
-                    sf::FloatRect visibleArea(0.f, 0.f, 
+                    sf::FloatRect visible_area(0.f, 0.f, 
                                              static_cast<float>(width), 
                                              static_cast<float>(height));
-                    window.setView(sf::View(visibleArea));
+                    window.setView(sf::View(visible_area));
                 }
                 simulator.handle_event(window, event);
             }
 
             // Update simulation
-            const float delta_time = delta_clock.restart().asSeconds();
+            const float DELTA_TIME = delta_clock.restart().asSeconds();
             simulator.update_mouse_position(window);
-            simulator.update(delta_time);
+            simulator.update(DELTA_TIME);
 
             // Render
-            window.clear(Theme::Background);
+            window.clear(Theme::BACKGROUND);
             simulator.render(window);
             window.display();
         }
